Allow returning no remote IP/port from API in some cases
It's not totally clear when this happens, but it does occasionally pop
up in real world interception, and results in missing events. Most
likely case I think is that the data is reset during some kinds of
errors, and so we miss it in those cases. Since it's rarely critical,
best to make it entirely optional I think. I'd love to prove the deep
cause to fix this fully but it's unlikely to be worthwhile at the mo.

I have done a bunch of testing here - best bet was that this was TLS
over H2 CONNECT, but it doesn't seem to repro in any tests I can
manage... Not worthwhile digging into this - just accept that sometimes
the remote address isn't available in certain edge cases.