Expose 'destination' on request data and avoid IPs in 'url' if possible
If you tunnel to a specific IP with a domain name specified, we'll use
the domain in the url property - this is normally clearer & preferable.
The IP is still available in the 'destination' property. It will still
be used in the URL if there's no hostname anywhere.

This also further deprecates the standalone 'hostname' field on
requests, which was not really used and functioned quite ambiguously.
You can now look at either the URL (best guess domain name), the
destination (the actual destination IP/domain name) or the Host/:Authority
header (the name the client is telling the server). In theory they
should probably match all the time, but they may not.

This mostly comes up when tunneling to one server, and sending a
different Host header, which mostly matters for weird edge cases and
testing of non-production deployments & similar.