Disable default request headers when proxying in modern Node (22.13+)
This is a breaking change as it changes the logic around message framing
when overriding with a custom body. In the past, we attempted to fix the
framing (i.e. setting a content-length header) any time you didn't
specifically disable this. Now, we _always_ fix the framing if you
modify a request or response body, to ensure it's always valid. This was
somewhat handled in the past by the default request headers, but now
that those are disabled the UX is very poor without this.

This should not affect you, unless you're intentionally setting an
invalid content-length or sending requests (not responses) with a body
but no framing (in which case they're likely unparseable anyway). Normal
valid requests will never be changed.