Handle framing automatically when downgrading proxied H2 traffic to H1
If you send an HTTP/2 request that likely has a body (we can't tell
initially) with no framing headers, we enable chunked encoding to ensure
any received body can be handled as expected. This might be subtly
breaking in some cases but should be semantically identical to the
existing behaviour for all valid requests.