Improve site cert generation further and drop node-forge entirely
This changes some specific details of the content of site certs to more
closely match standard browser TLS baseline requirements wherever
possible (currently: everything except AIA/CRL/OCSP which require public
CA URLs).

Note that this is a breaking change for generateSPKIFingerprint which is
now async and returns a Promise for a string, not a plain string.