Drop *HandlerDefinitions export to expose defns as *Handlers directly
Previously, {request,webSocket}Handlers was the implementation
instances while *HandlerDefinitions was the remote-client side handler
definitions themselves.

This is annoying though - you never need handler implementations
outside Mockttp, so in the few cases you need this, it just makes
all the names much longer for no benefit.

As step one of a larger redesign of this API, the *Handlers export now
exposes the definitions directly, and the *HandlerDefinitions export is
removed.