Scaffold support for multistep rules & rename all 'handlers' to 'steps'
This is a breaking change for the manual rule building APIs
(server.addRule({ handler: new XYZ })) but not for the most common &
standard chaining (server.forX().thenY()) rule definition APIs.

Internally this includes backward compatibility on both client & server
side, so should be safe to deploy either side independently, as long as
only one-step rules are used.

As yet there are no multi-step rules in place, so this is just
scaffolding & future planning for now (watch this space...)