Rename SimplePathMatcher & SimpleStep (previously SimpleHandler)
This is a breaking change only if you're manually building rule
structures yourself - there's no change to the chained rule building API
(forGet().thenReply(...))

These names sort-of made sense when we had few options, but no longer
really do. These are now called FlexiblePathMatcher (give it a path, it
will flexibly match it against the relative or absolute or protocol-less
URL automatically) and FixedResponseStep(Definition) which does exactly
what it says on the tin.