Make forwarding option a request transform, add path & query transforms
This is a breaking change if you're using thenForwardTo with a
beforeRequest callback (no longer allowed: return a url from
beforeRequest instead), or if you're manually building rules with a
'forwarding' option (use transformRequest.replaceHost instead).

In most cases though, this should work as before, but now supports
transformation of the path & query, match/replace transformation of the
host itself, and simplifies & aligns forwarding with other transform
options generally.