Allow importing CA keys in both PKCS#8 & PKCS#1 formats
Unfortunately it seems node-forge has been exporting to PKCS#1 up until
now, which is not well supported. We now export PKCS#8, but we still
want to handle all the existing keys out there, so we need to include
this as a fallback (and long-term too, it may be useful to support a
wider range of key inputs for custom Mockttp use cases).