Fix SOCKS handling of IPs with SNI passthrough & transforms
Previously, a SOCKS request to an IP (not DNS) resulted in that IP still
appearing in a few places, notably including transforms/callbacks &
upstream SNI when doing passthrough. This was a bug, as the effective
hostname (from Host header or inbound SNI) should be used instead, with
the IP only used for the actual TCP connection destination.