Fix client error timings further for multi-request connections
The previous fix handled simple cases OK, but assumed a request started
when the connection started, which isn't true for various cases. This is
starting to get overly precise about poorly defined details really, but
still - it's better if we use the time of the preceeding request as the
start, if there was one (either the start of this incomplete request, or
the previous request as a rough benchmark).