Fix HTTP/2 race condition with 204/304 responses + abort
In some cases (only sporadically with real traffic) it was possible to
see spurious aborts & inconsistent upstream/downstream data because the
HTTP/2 response would come in, we'd still be waiting for the stream to
close, but Node would have already successfully closed the downstream
stream for us. That resulted in a close event suggesting the client was
aborting before the server response completed, which wasn't true.

This isn't correct: if the writable has already ended then either it's
closed successfully (and we don't need the server response) or its
already aborted (and we still don't need the server response). Best to
close preemptively.