Enforce that rules must end or destroy before they resolve successfully
The one step breaking this invariant was the streaming step, which
resolved immediately. This meant it ignored any possible errors. With
this fix, stream errors will now be exposed in the response stream as
expected.

This also changes logic so that in general any handler throwing an
AbortError will have the underlying response automatically killed
if it hasn't destroyed it itself already. This shouldn't affect any
externally visible behaviour.