Update httpolyglot to fix unknown-over-TLS bug
This fixes a bug that this update causes. Previously, we depended on
subtle ordering of events to detect whether the last hop of a request
was encrypted or not (i.e. is a tunnelled request HTTPS or HTTP, given a
series of TLS or not hops en route).

This was buggy, and breaks with the latest httpolyglot, where this order
can change for requests without ALPN. Now, we properly set defaults &
reset encrypted status at each tunnel hop instead.

Also fixed an unrelated TS issue that breaks things with the latest Node
types in reuest-utils.