Fix client error events with unknown protocol & direct conns
Without this, direct conns would go into the unknown protocol handler if
they weren't HTTP, and would then get through to clientError if they
couldn't be forwarded, but we didn't include the normal client error
data so no useful debug info or context or error codes were provided.

We now expose useful error codes for these cases, and include all the
raw data required to typically provide a bit of a glimpse of the raw
data that's being sent here.

This shouldn't really come up - almost all unknown protocol handling
should really use SOCKS or CONNECT tunnels to avoid this - but when it
does it's good to get proper info.