Fix raw passthrough containing TLS
We now have a confusing scenario with TLS passthrough (non-intercepted)
raw passthrough (fully intercepted) and raw-passthrough-with-tls (fully
intercepted, but we unwrap and then recreate TLS so we can see
everything).

Previously the last case there didn't work: we unwrapped TLS, then
forwarded the traffic it contained to the destination without TLS. Bad
for security, but also in practice extremely likely to not work at all.

We now recreate the TLS connection on the passthrough, so
unknown-over-TLS should work as expected with TLS on both sides but
traffic still inspectable (cool right). This allows us to do things like
secure MQTT, at least in theory (in practice it's unclear how tricky
cert trust will be in scenarios like that).